have diff parent classes with or without JWT
